class library: NodeProxy <<> shouldn't reshape argument
The method was written under the assumption that the argument (the proxy to which the receiver's control is mapped) is static, and not elastic. This commit makes sure it stays the same. Since any rate - numChannel combination can be mapped using the "/n_set" message, any such differences are taken care for by the backend.