Fix race condition in scheduler
When running `.sc` files with `sclang`, the execution of the of a
SuperCollider file containing `0.exit` can complete quickly causing
`schedStop()` to be called (on the main thread) _before_
`schedRunFunc()` (scheduler thread) acquires the `gLangMutex` lock for
the first time. This can result bugs fromin a race condition causing
`sclang` to hand indefinitely upon exiting.

This commit removes this race condition from the code with two changes:

The `schedRunFunc()` has been changed to check the `gRunSched`
immediately after acquiring `gLangMutex` for the first time. This makes
sure `schedRunFunc()` always returns even if `stopSched()` is called
before the scheduler thread can initially acquire the mutex.

The line that sets `gRunSched` to `true` initially has been moved from
`schedRunFunc()` on scheduler thread to be in `schedRun()` on the main
thread. The guarantees that `stopSched()` will observe `gRunSched` as
true, because the same thread that calls `stopSched()` also sets
`gRunSched` to true: the main thread.