Servers: workaround for boost interprocess (#5393)
By default, Boost interprocess was using a system log message to choose the name for the shared memory directory. This broke sometimes when people's computers would not have those log messages present. Boost interprocess provides a second way of achieving the same thing, which we tested and confirmed to work on one windows machine.